feature : saving parameters in progress in the settings menu [WIP]

+++ Assets/03_Scripts/02_GameOfLifeLevel/SaveDataGridParameters.cs
@@ -9,7 +9,7 @@ public class SaveDataGridParameters
 
     public float musicVolume;
 
-    public int effectVolume;
+    public float effectVolume;
 
-    public int uiVolume;
+    public float uiVolume;
 }
